%<+% {ggfun}	R Documentation

%<+%

Description

This operator attaches annotation data to a ggtree or ggsc graphic object

Usage

p %<+% data

Arguments

p
ggplot2 object, such as ggtree or ggsc graphic object.

data
data.frame, which must contains a column of node, or the first column of taxa labels, when p is a ggtree object. Or it must contains columns of .BarcodeID, when p is a ggsc object and p$data does not contain a column of features, if it contains, the data must also contains a column of features.

Value

ggplot object with annotation data added

[Package ggfun version 0.1.5 Index]