facet_set {ggfun}	R Documentation

facet_set

Description

add a facet label to a ggplot or change facet label of a ggplot

Usage

facet_set(label, side = "t", angle = NULL)

Arguments

label
a character or a named vector to label the plot

side
to label the plot at which side, either 't' (top) or 'r' (right)

angle
angle of the facet label. Default is 0 for side='t' and -90 for side='r'.

Value

a ggplot with facet label

[Package ggfun version 0.1.5 Index]